Make CommandsModule own its parsing and accumulated state
Previously CommandsModule took an opaque CommandsHandler callback and
plugin.go held the commands map plus the closure that called
commandsFromTable. That left plugin.go knowing about the full Lua-side
command parsing surface (commandsFromTable, userCommandFromLua, the
luaStringField/luaBoolField/luaStringSliceField/luaExitField helpers),
which defeated the point of extracting modules.

Move the commands map onto CommandsModule and relocate every command
parsing helper from plugin.go into module_commands.go. plugin.go now
constructs a fresh CommandsModule per Init, hands it to NewRuntime, and
reads cmdModule.Commands() out only on success — the partial-init
guarantee is preserved and plugin.go drops to ~85 lines of pure
lifecycle.

CommandsHandler is gone (the closure indirection it enabled is no
longer needed).
